Struct armorlib::binary_object::BinaryObject

pub struct BinaryObject {
    pub file_path: Option<String>,
    pub data: Vec<u8>,
}

A struct that represents an unstructured piece of binary data in the form of a Vec<u8>. It also includes the option of having a String to designate the path.

Fields
file_path: Option<String>

The path on the host operating system where this BinaryObject resides. Because this field is not always applicible (for example, binary objects which exist only in a database), the possibility of absence is represented in the Option.

data: Vec<u8>

A Vec<u8> of the binary object's data.