Module armorlib::finding

This module defines the Finding, a struct that represents a vulnerability found by a ScanModule.

Structs

Finding
A struct that represents a vulernability found by a ScanModule.

Enums

Severity
An enum that represents severity. While not necessarily limited to findings, they are commonly used in the Finding struct to designate the severity of a finding related to a privacy or security vulnerability. The following documentation assumes this use case.