Enum armorlib::finding::Severity

pub enum Severity {
    Ok(String),
    Warn(String),
    Danger(String),
    Severe(String),
}

An enum that represents severity. While not necessarily limited to findings, they are commonly used in the Finding struct to designate the severity of a finding related to a privacy or security vulnerability. The following documentation assumes this use case.

The attached String is a description of the potential ramifications of the vulnerability. For example, this explanation string may be likely indicates malicious software or, alternatively, not an issue. This should not be a full sentence.

Variants
Ok(String)

For use there is no issue.

Warn(String)

For use when the data has a potential problem, but it is not threatening; an afterthought.

Danger(String)

For use when the data has a problem which may compromise security or privacy.

Severe(String)

For use then the data has a problem which will surely compromise security or privacy.