Struct armorlib::finding::Finding

pub struct Finding {
    pub title: String,
    pub id: String,
    pub description: String,
    pub severity: Severity,
}

A struct that represents a vulernability found by a ScanModule.

Fields
title: String

The title of the vulenrability. This is human-meaningful, and should make sense on its own with only the context of the scan module's name. This shouldn't be a complete sentence; the knowledge of the scan module should make only a noun necessary; that is, the title indicates what the scan module actually found. Therefore, it is not necessary to include a verb.

For example, the title of a Finding created by the strings scan module might be XOR DECODE LOOP (COMMON TO MICROSOFT MALWARE). Because this will be listed explicitly as a finding of the Strings scan module, such a title will make sense to the user.

id: String

A consistent and unique identifier for the type of finding in screaming snake case. All findings of the same type should have the same ID. For example, the ID of a finding of the string scan module might be SUSPICIOUS_STRING.

description: String

A human meaningful description of the finding. This should be between one and two full sentences.

severity: Severity

The Severity of the finding. This ranges from Ok to Severe. See the documentation for Severity for more in-depth explanations of each possible state.