Function armorlib::preprocessors::process

pub fn process(
    preprocessors: Vec<Box<Preprocessor>>, 
    binary_object: &BinaryObject
) -> HashMap<String, HashMap<String, String>>

Process the given Vec<Box<Preprocessor>> on the given &BinaryObject and return a HashMap<String, HashMap<String>>. While concurrency is not yet available in ArmorLib, it will be implemented in this function, if anywhere.

In nearly all cases, it is better to perform File.process(), Vec<u8>.process(), or even coordinator::process() than to use this function. The previous functions will make sure that everything is set up properly; using this function alone will require you to manage the preprocessors and later scanning yourself. If you need fine grained control, use this function. Otherwise, don't—you'll save yourself a headache.

Arguments

preprocessors: a Vec<Box<preprocessors>> of the preprocessors to be run.

binary_object: a reference to a BinaryObject on which the preprocessors will run.