Trait armorlib::process::Process

pub trait Process {
    fn process(self) -> Result<ScanResult, ArmorlibError>
    where
        Self: Sized,
        BinaryObject: From<Self>,
    { ... }
}

A trait that allows for the object to be run through the ArmorLib system with only a single call to .process(). This is a perfectly valid way of passing objects through ArmorLib, however coordinator::Process is also available for more granular control.

Examples

use armorlib::Process;
let data: Vec<u8> = vec![]; // empty data for demo
let scan_result = data.process().unwrap();

Provided Methods

fn process(self) -> Result<ScanResult, ArmorlibError> where
Self: Sized,
BinaryObject: From<Self>,