Module armorlib::scan_module

This module defines the ScanModule, a trait that all scan modules must implement in order to be run by ArmorLib. The ScanModule trait is available in the root namespace as armorlib::ScanModule.

Traits

ScanModule
A trait that defines the necessary functions for ScanModules to implement. To contribute a new scan module, it will also need to be included in mod.rs inside the scan_modules directory. A scan module is a modular component that finds vulnerabilities to privacy and security inside ScanObjects. The ScanModule is the core component of ArmorLib.